requirements.txt
langgraph>=0.0.15
boto3>=1.34.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
pydantic>=2.5.0
typing-extensions>=4.8.0 